Rename extension
In the beggining of 2018 this extension was taken down from Chrome Web
Store:

"Google was notified that some of your materials allegedly infringe upon
the trademarks of others, the details of the removed extension may be
found at the end of this message."

I believe this was because it was called Facebook Incognito and Facebook
normally calls their products "Facebook something", for example Facebook
Messenger.

I replied to the takedown notice asking if a rename to "Incognito for
Facebook" would be ok but never received a word back. Thanks.